Move default Planning class from "core.py" to separate module
* Add NonClassicAuctionPlannig class (from auctions.insider)
