+++ openprocurement/auction/includeme.py
@@ -2,8 +2,9 @@
 from zope.interface import Interface
 from zope.interface.interface import InterfaceClass
 
-from openprocurement.auction.core import RunDispatcher, Planning
+from openprocurement.auction.core import RunDispatcher
 from openprocurement.auction.interfaces import IFeedItem, IAuctionDatabridge, IAuctionsChronograph
+from openprocurement.auction.plannings import ClassicAuctionPlanning
 from openprocurement.auction.utils import get_logger_for_calling_module
 
 
@@ -12,7 +13,7 @@ def _register(components, procurement_method_type):
                            bases=(Interface,))
     components.add_auction(iface,
                            procurementMethodType=procurement_method_type)
-    components.registerAdapter(Planning, (IAuctionDatabridge, IFeedItem), iface)
+    components.registerAdapter(ClassicAuctionPlanning, (IAuctionDatabridge, IFeedItem), iface)
     components.registerAdapter(RunDispatcher, (IAuctionsChronograph, IFeedItem), iface)
 
     # for proper plugin loading logs
